refactor(agent): extract composer picker hook
