+++ docs/progress/NEXT_LOW_HANGING_CLEANUPS.md
@@ -12,8 +12,8 @@ Process: implement one task at a time, run focused checks, get subagent review w
 | NLH-02 | done | Extract `CommandPalette` recent/select handlers into a focused helper/hook. | reviewer: ship | `88bb454e` |
 | NLH-03 | done | Move `convertBlobUrlToDataUrl` into a browser file utility while preserving prompt-input re-export. | reviewer: ship | `d6626992` |
 | NLH-04 | done | Split `prompt-input.tsx` constants/context only, with no JSX movement. | reviewer: ship after one revise round | `cc3d85ca` |
-| NLH-05 | done | Extract ChatPanel composer history hook. | reviewer: ship | pending |
-| NLH-06 | pending | Extract ChatPanel slash/mention handlers. | pending | pending |
+| NLH-05 | done | Extract ChatPanel composer history hook. | reviewer: ship | `6d79f49b` |
+| NLH-06 | done | Extract ChatPanel slash/mention handlers. | reviewer: ship | pending |
 
 ## Notes
